+++ src/test/converter2/behavior/overloads.ts
@@ -1,13 +1,14 @@
+/**
+ * No arg comment
+ * {@label NO_ARGS}
+ */
 export function foo(): string;
 /**
- * Overrides summary
- */
-export function foo(x: string): number;
-
-/**
- * Implementation comment
+ * {@inheritDoc (foo:NO_ARGS)}
+ * {@label WITH_X}
  * @param x docs for x
  */
+export function foo(x: string): number;
 export function foo(x?: string): string | number {
     return x == null ? "" : x.length;
 }